backend/requirements.txt
Flask
Flask-Cors
mysql-connector-python
Werkzeug
requests
python-dotenv